Refactor share space
